# docs/swagger.yaml
basePath: /api/v1
definitions:
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  jsontype.Tags:
    additionalProperties: {}
    type: object
...
  omc.Device:
    properties:
      ActiveStatus:
        type: string
      Created:
        type: string
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      Enable:
        type: boolean
      Group:
        $ref: '#/definitions/omc.Group'
      GroupId:
        type: integer
      Id:
        type: integer
      LastInformTime:
        type: string
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      Methods:
        items:
          type: string
        type: array
      Name:
        type: string
      Online:
        type: boolean
      Oui:
        type: string
      ParameterNotifications:
        $ref: '#/definitions/omc.ParameterNotifications'
      ParameterValues:
        $ref: '#/definitions/omc.ParameterValues'
      ParameterWritables:
        $ref: '#/definitions/omc.ParameterWritables'
      Product:
        $ref: '#/definitions/omc.Product'
      ProductClass:
        type: string
      ProductId:
        type: integer
      ProductType:
        type: string
      Properties:
        $ref: '#/definitions/jsontype.Tags'
      Schema:
        type: string
      SerialNumber:
        type: string
      Updated:
        type: string
    type: object
...
  omc.DeviceEvent:
    properties:
      CurrentTime:
        type: string
      Device:
        $ref: '#/definitions/omc.Device'
      DeviceId:
        type: integer
      EventType:
        type: string
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      Oui:
        type: string
      ProductClass:
        type: string
      ProductType:
        type: string
      Schema:
        type: string
      SerialNumber:
        type: string
      Time:
        type: string
      Updated:
        type: string
    type: object
...
  omc.Firmware:
    properties:
      Created:
        type: string
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      Id:
        type: integer
      Name:
        type: string
      ProductType:
        type: string
      S3Object:
        $ref: '#/definitions/s3.S3Object'
      S3ObjectId:
        type: integer
      Schema:
        type: string
      Updated:
        type: string
      UploadTime:
        type: string
      Uploader:
        type: string
      Version:
        type: string
      products:
        items:
          $ref: '#/definitions/omc.Product'
        type: array
    type: object
  omc.Group:
    properties:
      Children:
        items:
          $ref: '#/definitions/omc.Group'
        type: array
      Created:
        type: string
      Id:
        type: integer
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      Name:
        type: string
      Parent:
        $ref: '#/definitions/omc.Group'
      ParentId:
        type: integer
      Schema:
        type: string
      Updated:
        type: string
    type: object
...
  omc.ParameterNotifications:
    additionalProperties:
      type: integer
    type: object
  omc.ParameterValues:
    additionalProperties:
      type: string
    type: object
  omc.ParameterWritables:
    additionalProperties:
      type: boolean
    type: object
  omc.Product:
    properties:
      Created:
        type: string
      Default:
        type: boolean
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      Enable:
        type: boolean
      Firmwares:
        items:
          $ref: '#/definitions/omc.Firmware'
        type: array
      Id:
        type: integer
      Manufacturer:
        type: string
      MetaData:
        $ref: '#/definitions/jsontype.Tags'
      ModelName:
        type: string
      Oui:
        type: string
      ParameterPath:
        type: string
      PerformanceValueDefines:
        $ref: '#/definitions/jsontype.Tags'
      ProductClass:
        type: string
      ProductType:
        type: string
      Schema:
        type: string
      SupportedAlarms:
        items:
          $ref: '#/definitions/omc.ProductSupportedAlarm'
        type: array
      Updated:
        type: string
    type: object
  omc.ProductSupportedAlarm:
    properties:
      AlarmIdentifier:
        type: string
      Created:
        type: string
      EventType:
        type: string
      Id:
        type: integer
      Oui:
        type: string
      PerceivedSeverity:
        type: string
      ProbableCause:
        type: string
      Product:
        $ref: '#/definitions/omc.Product'
      ProductClass:
        type: string
      ProductId:
        type: integer
      ProductType:
        type: string
      Schema:
        type: string
      SerialNumber:
        type: string
      SpecificProblem:
        type: string
      Updated:
        type: string
    type: object
...
  omc.listDeviceEventsData:
    properties:
      Data:
        items:
          $ref: '#/definitions/omc.DeviceEvent'
        type: array
      Total:
        type: integer
    type: object
...
  omc.listFirmwaresData:
    properties:
      Data:
        items:
          $ref: '#/definitions/omc.Firmware'
        type: array
      Total:
        type: integer
    type: object
...
  omc.setFirmwareProductsBody:
    properties:
      ModelNames:
        items:
          type: string
        type: array
    type: object
...
  omc.updateDeviceInfoBody:
    properties:
      Name:
        type: string
    type: object
...
  s3.S3Object:
    properties:
      Bucket:
        type: string
      Created:
        type: string
      Deleted:
        $ref: '#/definitions/gorm.DeletedAt'
      DownloadUrl:
        type: string
      FileName:
        type: string
      FileSize:
        type: integer
      Id:
        type: integer
      Key:
        type: string
      MetaData:
        $ref: '#/definitions/s3.Tags'
      Schema:
        type: string
      SignMethod:
        type: string
      Signature:
        type: string
      Updated:
        type: string
    type: object
  s3.Tags:
    additionalProperties: {}
    type: object
...
host: dev.netdoop.com
info:
  contact: {}
  title: NetDoop API
  version: "1.0"
paths:
...
  /omc/device-events:
    get:
      operationId: list-device-events
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - default: ""
        description: Sort order
        in: query
        name: order_by
        type: string
      - default: ""
        description: Query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total:
              description: Total number
              type: int
          schema:
            $ref: '#/definitions/omc.listDeviceEventsData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: List device events
      tags:
      - OMC DeviceEvents
  /omc/device-events/{ts}:
    get:
      operationId: get-device-event
      parameters:
      - description: Timestamp
        in: path
        name: ts
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.DeviceEvent'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Get device event
      tags:
      - OMC DeviceEvents
...
  /omc/devices/{id}:
    delete:
      consumes:
      - application/json
      operationId: delete-device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Delete a device
      tags:
      - OMC Devices
    get:
      consumes:
      - application/json
      operationId: get-device
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.Device'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Get a device
      tags:
      - OMC Devices
    put:
      consumes:
      - application/json
      operationId: update-device-info
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: integer
      - description: Device
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/omc.updateDeviceInfoBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.Device'
        "400":
          description: 'Bad Request: invalid input parameter'
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Update device info
      tags:
      - OMC Devices
...
  /omc/devices/{id}/methods:
    get:
      consumes:
      - application/json
      operationId: list-device-methods
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              type: string
            type: array
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Get the methods of a device
      tags:
      - OMC Devices
  /omc/devices/{id}/parameters:
    get:
      consumes:
      - application/json
      operationId: list-device-parameters
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.ParameterValues'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Get the parameters of a device
      tags:
      - OMC Devices
...
  /omc/firmwares:
    get:
      consumes:
      - application/json
      operationId: list-firmwares
      parameters:
      - default: 1
        description: Page
        in: query
        name: page
        type: integer
      - default: 20
        description: Page size
        in: query
        name: page_size
        type: integer
      - default: ""
        description: Sort order
        in: query
        name: order_by
        type: string
      - default: ""
        description: Query
        in: query
        name: q
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          headers:
            X-Total:
              description: Total number
              type: int
          schema:
            $ref: '#/definitions/omc.listFirmwaresData'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: List firmwares
      tags:
      - OMC Firmwares
    post:
      consumes:
      - multipart/form-data
      operationId: create-firmware
      parameters:
      - description: Version
        in: formData
        name: Version
        required: true
        type: string
      - description: ProductType
        in: formData
        name: ProductType
        required: true
        type: string
      - description: Product list
        in: formData
        name: Products
        required: true
        type: string
      - description: Firmware file to upload
        in: formData
        name: File
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.Firmware'
        "400":
          description: 'Bad Request: invalid input parameter'
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Create firmware
      tags:
      - OMC Firmwares
  /omc/firmwares/{id}:
    delete:
      consumes:
      - application/json
      operationId: delete-firmware
      parameters:
      - description: Firmware ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Delete firmware
      tags:
      - OMC Firmwares
    get:
      operationId: get-firmware
      parameters:
      - description: Firmware ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.Firmware'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal Server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Get firmware
      tags:
      - OMC Firmwares
  /omc/firmwares/{id}/products:
    put:
      consumes:
      - application/json
      operationId: set-firmware-products
      parameters:
      - description: Device ID
        in: path
        name: id
        required: true
        type: integer
      - description: Body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/omc.setFirmwareProductsBody'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/omc.Device'
        "400":
          description: 'Bad Request: invalid input parameter'
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      security:
      - Bearer: []
      summary: Set firmware products
      tags:
      - OMC Firmwares
...
schemes:
- http
securityDefinitions:
  Bearer:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"